Re-work of codebase to improve separation of concerns.
Focused on ports mainly as all port functionality was housed in MMALPortBase/MMALPortImpl which didn't make sense. There is now a proper representation of Input/Output/Control ports and have been interfaced accordingly. Capture handlers have been moved to port level as it feels better for them to be associated with the ports processing data rather than the component as a whole. More work on tests - there seems to be some strange behaviour when transitioning between Image/Configuration/Video test classes, possibly to do with the Fixture, this needs investigating.